+++ ecs-20140526/core/Properties/AssemblyInfo.cs
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("2a412ba0-60d5-11f0-ae1c-638892d6de88")]
+[assembly: Guid("f3f21320-63f9-11f0-979c-b390e95aeb1e")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("7.0.13.0")]
-[assembly: AssemblyFileVersion("7.0.13.0")]
+[assembly: AssemblyVersion("7.0.14.0")]
+[assembly: AssemblyFileVersion("7.0.14.0")]